video trailer haleko

+++ lib/elements/music.dart
@@ -2,8 +2,9 @@ class Music {
   final int id;
   final String artist_name;
   final String track_name;
+  //final List<String> genre;
   final String genre;
-  final double release_date;
+  final int release_date;
 
   Music({
     required this.id,
@@ -14,10 +15,23 @@ class Music {
   });
 
   factory Music.fromJson(Map<String, dynamic> json) {
+    // List<String> genreList = (json['genre'] as List).cast<String>();
+    // List<String> genreList = [];
+
+    // if (json['genre'] is List) {
+    //   // Use the genre list as is
+    //   genreList = (json['genre'] as List).map((e) => e.toString()).toList();
+    // } else if (json['genre'] is String) {
+    //   // Parse the genre string into a List<String>
+    //   var rawGenreList = json['genre'].replaceAll("[", "").replaceAll("]", "");
+    //   genreList = rawGenreList.split(',').map((e) => e.trim()).toList();
+    // }
+
     return Music(
       id: json['id'],
       artist_name: json['artist_name'],
       track_name: json['track_name'],
+      //genre: genreList,
       genre: json['genre'],
       release_date: json['release_date'],
     );
